backend/requirements.txt
flask
flask-socketio
python-bitcoinrpc